requirements.txt
fastapi
python-dotenv
streamlit
pymupdf
openai
plotly
matplotlib
seaborn
uv
uvicorn
fastmcp